diffuse as module
